pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soil-resnet-model"
version = "1.0.0"
description = "NationalSoilNet: Multi-Target Soil Property Prediction from Satellite Embeddings"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jack Coombs"},
]
keywords = ["soil", "deep-learning", "remote-sensing", "pytorch", "resnet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "rasterio>=1.3",
    "geopandas>=0.14",
    "shapely>=2.0",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
baselines = [
    "catboost>=1.2",
    "cubist>=0.2",
]
gee = [
    "earthengine-api>=0.1.370",
    "geemap>=0.30",
]
notebooks = [
    "jupyter>=1.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
]
data = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
]
all = [
    "soil-resnet-model[baselines,gee,notebooks,data,dev]",
]

[project.urls]
Repository = "https://github.com/jackcoombs/soil-resnet-model"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
